+++ Sources/BlueTriangle/CrashReportManager.swift
@@ -42,6 +42,14 @@ final class CrashReportManager: CrashReportManaging {
             self?.startupTask = nil
         }
     }
+    
+    func stop(){
+        //Clear saved crash
+        self.startupTask?.cancel()
+        self.startupTask = nil
+        CrashReportPersistence.disableExaptionHandler()
+        crashReportPersistence.clear()
+    }
 
     func uploadCrashReport(session: Session) {
         guard let crashReport = crashReportPersistence.read() else {
@@ -67,12 +75,15 @@ final class CrashReportManager: CrashReportManaging {
         function: StaticString,
         line: UInt
     ) {
+        guard let session = session() else {
+            return
+        }
         
         let nativeApp = NativeAppProperties.nstEmpty
         let report = ErrorReport(nativeApp: nativeApp, eTp: BT_ErrorType.NativeAppCrash.rawValue, error: error, line: line, time: intervalProvider().milliseconds)
         let pageName = BlueTriangle.recentTimer()?.page.pageName
         do {
-            try upload(session:session() , report: report, pageName: pageName)
+            try upload(session:session , report: report, pageName: pageName)
         } catch {
             logger.error(error.localizedDescription)
         }
@@ -134,4 +145,5 @@ private extension CrashReportManager {
                                                        report: report, pageName: pageName)
         uploader.send(request: reportRequest)
     }
+    
 }