+++ Sources/BlueTriangle/Protocols/CrashReportManaging.swift
@@ -10,4 +10,5 @@ import Foundation
 protocol CrashReportManaging {
     func uploadCrashReport(session: Session)
     func uploadError<E: Error>(_ error: E, file: StaticString, function: StaticString, line: UInt)
+    func stop()
 }